==> 1_EDA_Prep.ipynb <==
# Auto Insurance Claim Fraud Indicators and Classification

# Mary Donovan Martello

## The goal of this project was to identify significant features in fraudulent insurance claim transactions and to design predictive classification models to predict whether fraud was reported on the insurance claim transaction.  This notebook includes exploratory data analysis and data preparation for the project.

# Part 1:  Exploratory Data Analysis and Data Preparation

# import data set and libraries for data preparation phase
import pandas as pd
import numpy as np
# importing regex module (search strings) RegEx can be used to check if a string contains the specified search pattern
import re


### The dataset includes 1,000 prior claim transaction records.  Each record has a mix of 38 quantitative and categorical data features about the claim filed, including information on the policy, insured, and automobile, aspects of the damage incident, and elements of the claim filed.  The dataset also has a feature that indicates whether fraud was reported on each observation (i.e., either Y or N).

dfClaims = pd.read_csv('FradulentInsuranceClaims.csv')

dfClaims.head(2)

### Data Cleaning

# print the column names
dfClaims.columns

# rename difficult column names
dfClaims.rename(columns={'capital-gains': 'capital_gains', 'capital-loss': 'capital_loss'}, inplace=True)
dfClaims.columns

# check whether the data has any null values 

# There are '?' in the datset which are removed by NaN Values
dfClaims = dfClaims.replace('?',np.NaN)

dfClaims.isnull().any()



# check df shape
dfClaims.shape

# check which rows have null values
is_NaN = dfClaims.isnull()
row_has_NaN = is_NaN.any(axis=1)
rows_with_NaN = dfClaims[row_has_NaN]

#print(rows_with_NaN)

# There are '?" in three features; address these using fillna

# replace the '?' by the most common collision type 
dfClaims['collision_type'].fillna(dfClaims['collision_type'].mode()[0], inplace = True)

# property damage is either yes or no; fill with no because non responses for property damage might mean no property damage.
dfClaims['property_damage'].fillna('NO', inplace = True)

# again, if there are no responses for police report available it may mean No report available
dfClaims['police_report_available'].fillna('NO', inplace = True)

dfClaims.isnull().any().any()


# which features have null values?
dfClaims.isnull().any()

# removing null values to avoid errors: the _c39 feature is all null values
dfClaims.dropna(axis='columns', thresh=100, inplace = True) 

# confirm that all null values have been removed
dfClaims.isnull().any().any()

# confirm ?s replaced:  were 177 ?
dfClaims['collision_type'].value_counts()


# confirm ?s replaced:  were 360 ?
dfClaims['property_damage'].value_counts()


# confirm ?s replaced:  were 342 ?
dfClaims['police_report_available'].value_counts()


# fraud counts
dfClaims.fraud_reported.value_counts().sort_index()

# split policy_csl into separate columns as it has two values per column

# new data frame with split value columns 
newDF = dfClaims["policy_csl"].str.split("/", n = 1, expand = True)

# making separate csl_bodily column from new data frame 
dfClaims["csl_bodily"] = newDF[0] 
  
# making separate csl_property column from new data frame 
dfClaims["csl_prop"] = newDF[1] 
  
# Dropping old policy_csl columns 
dfClaims.drop(columns =["policy_csl"], inplace = True) 

# convert from string to int
dfClaims["csl_bodily"] = dfClaims["csl_bodily"].astype(int)
dfClaims["csl_prop"] = dfClaims["csl_prop"].astype(int)

# df display 
dfClaims.columns

# check shape again
dfClaims.shape

# drop umbrella outlier observation
dfClaims.drop(dfClaims[dfClaims['umbrella_limit'] < 0].index, inplace = True)
dfClaims.shape

**Create derived feature for length of time claimant was a customer before claim incident.**

# overwriting data after changing date format 
dfClaims["incident_date"]= pd.to_datetime(dfClaims["incident_date"])
dfClaims["policy_bind_date"]= pd.to_datetime(dfClaims["policy_bind_date"])
dfClaims.info()

# create new column for diff in months between policy_bind_date and incident_date
dfClaims['months_bf_incident'] = ((dfClaims.incident_date - dfClaims.policy_bind_date)/np.timedelta64(1, 'M'))
dfClaims['months_bf_incident'].astype(int)

# Dropping old date coluns 
dfClaims.drop(columns =["policy_bind_date", "incident_date"], inplace = True)

# confirm columns
dfClaims.columns

# confirm shape
dfClaims.shape

## Data Understanding

### This project utilized the EDA from this Exploratory Data Analysis and Hypothesis Testing project: [EDA.](https://github.com/MaryDonovanMartello/EDA-and-Hypothesis-Testing)

# Create new dataset with quantitative variables (exlcluding ones with too many nonunique values to be helpful)
dfClaims_quantVars = dfClaims.filter(['months_as_customer', 'age', 'policy_deductable', 'policy_annual_premium', 'umbrella_limit', 'capital_gains', 'capital_loss', 'incident_hour_of_the_day', 'number_of_vehicles_involved', 'bodily_injuries', 'witnesses', 'total_claim_amount', 'injury_claim', 'property_claim', 'vehicle_claim', 'auto_year', 'csl_bodily', 'csl_prop'], axis=1)


dfClaims_quantVars.shape

### Visuals

**This project utilized visuals from the 2_R_Visuals file and the EDA from this Exploratory Data Analysis and Hypothesis Testing project:*** [EDA.](https://github.com/MaryDonovanMartello/EDA-and-Hypothesis-Testing)

### Descriptive Statistics

# import libraries for statistical exploratory analysis
from collections import Counter
import random
import scipy.stats
import statsmodels.formula.api as smf
from scipy import stats
import patsy #stats
from tabulate import tabulate

**Each record in the dataset includes a label of whether the claim was classified as fraudulent (fraud = Y) or as non-fraudulent (fraud = N). Create separate subsets of the dataset so that one subset only has records that did not have fraud transactions and another subset that only has records with fraud transactions. Use these separate subsets for EDA.**

# subset for fraud reported if yes
rowsYesFraud = dfClaims['fraud_reported'] == 'Y'
fraudY2 = dfClaims.loc[rowsYesFraud, ]

# subset for fraud reported is no
rowsNoFraud = dfClaims['fraud_reported'] == 'N'
fraudN2 = dfClaims.loc[rowsNoFraud, ]

# Create new dataset with quantitative variables - full dataset
dfClaims_quant = dfClaims.filter(['months_bf_incident', 'total_claim_amount', 'witnesses', 'policy_deductable', 'umbrella_limit'], axis=1)
dfClaims_quant.describe()

# Create new dataset with quantitative variables - fraud reported
fraudY2_quant = fraudY2.filter(['months_bf_incident', 'total_claim_amount', 'vehicle_claim', 'witnesses', 'number_of_vehicles_involved', 'policy_deductable', 'umbrella_limit'], axis=1)
fraudY2_quant.describe()

# Create new dataset with quantitative variables - no fraud reported
fraudN2_quant = fraudN2.filter(['months_bf_incident', 'total_claim_amount', 'vehicle_claim', 'witnesses', 'number_of_vehicles_involved', 'policy_deductable', 'umbrella_limit'], axis=1)
fraudN2_quant.describe()

**Correlation of quantitative features.**

# Create new dataset with quantitative variables - full dataframe
dfClaims_quant = dfClaims.filter(['months_bf_incident', 'total_claim_amount', 'witnesses', 'policy_deductable', 'umbrella_limit'], axis=1)
dfClaims_quant.corr()

# Create new dataset with quantitative variables - fraud reported
fraudY2_quant = fraudY2.filter(['months_bf_incident', 'total_claim_amount', 'witnesses', 'policy_deductable', 'umbrella_limit'], axis=1)
fraudY2_quant.corr()

# Create new dataset with quantitative variables - no fraud reported
fraudN2_quant = fraudN2.filter(['months_bf_incident', 'total_claim_amount', 'witnesses', 'policy_deductable', 'umbrella_limit'], axis=1)
fraudN2_quant.corr()

**Point-biserial correlation is used for correlation between the binary fraud reported variable and continuous variables.  None of the point-biserial correlations for the fraud_reported feature are strong.  However, the p-value for umbrella limit shows that it may be significant.**

# factorize categorical from strings to numberics in dfClaims df
dfClaims['fraud_reported_f'] = pd.factorize(dfClaims.fraud_reported)[0]

pbcClaimAmt = stats.pointbiserialr(dfClaims.fraud_reported_f, dfClaims.total_claim_amount)
print(pbcClaimAmt)

pbcUmb = stats.pointbiserialr(dfClaims.fraud_reported_f, dfClaims.umbrella_limit)
print(pbcUmb)

pbcMonths = stats.pointbiserialr(dfClaims.fraud_reported_f, dfClaims.months_as_customer)
print(pbcMonths)

pbcWitn = stats.pointbiserialr(dfClaims.fraud_reported_f, dfClaims.witnesses)
print(pbcWitn)

pbcDeduc = stats.pointbiserialr(dfClaims.fraud_reported_f, dfClaims.policy_deductable)
print(pbcDeduc)

# export the cleaned df to a csv file
dfClaims.to_csv('dfClaims.csv', index = False)

## Data Preparation

### Address skew in quantitative features with log-transformation.

# log-transformation
# first need to convert values in capital_loss to remove negative values before log_transformation.  
#     Add the largest negative number to all values.

# largest negative value
largestNeg = dfClaims['capital_loss'].min()

#dfClaims_quantVars['capital_loss'] = dfClaims_quantVars['capital_loss'].apply(lambda x: x + 1 - min(x))
dfClaims_quantVars['capital_loss'] = dfClaims_quantVars['capital_loss'] - largestNeg

dfClaims_quantVars['capital_loss'].min()

# log-transformation - cont

# now without negative values, can use log_transformation

def log_transformation(data):
    return data.apply(np.log1p)

dfClaims_quantVars['months_as_customer'] = log_transformation(dfClaims_quantVars['months_as_customer'])
dfClaims_quantVars['age'] = log_transformation(dfClaims_quantVars['age'])
dfClaims_quantVars['policy_deductable'] = log_transformation(dfClaims_quantVars['policy_deductable'])
dfClaims_quantVars['policy_annual_premium'] = log_transformation(dfClaims_quantVars['policy_annual_premium'])
dfClaims_quantVars['umbrella_limit'] = log_transformation(dfClaims_quantVars['umbrella_limit'])
dfClaims_quantVars['capital_gains'] = log_transformation(dfClaims_quantVars['capital_gains'])
dfClaims_quantVars['capital_loss'] = log_transformation(dfClaims_quantVars['capital_loss'])
dfClaims_quantVars['incident_hour_of_the_day'] = log_transformation(dfClaims_quantVars['incident_hour_of_the_day'])
dfClaims_quantVars['number_of_vehicles_involved'] = log_transformation(dfClaims_quantVars['number_of_vehicles_involved'])
dfClaims_quantVars['bodily_injuries'] = log_transformation(dfClaims_quantVars['bodily_injuries'])
dfClaims_quantVars['witnesses'] = log_transformation(dfClaims_quantVars['witnesses'])
dfClaims_quantVars['total_claim_amount'] = log_transformation(dfClaims_quantVars['total_claim_amount'])
dfClaims_quantVars['injury_claim'] = log_transformation(dfClaims_quantVars['injury_claim'])
dfClaims_quantVars['property_claim'] = log_transformation(dfClaims_quantVars['property_claim'])
dfClaims_quantVars['vehicle_claim'] = log_transformation(dfClaims_quantVars['vehicle_claim'])
dfClaims_quantVars['auto_year'] = log_transformation(dfClaims_quantVars['auto_year'])
dfClaims_quantVars['csl_bodily'] = log_transformation(dfClaims_quantVars['csl_bodily'])
dfClaims_quantVars['csl_prop'] = log_transformation(dfClaims_quantVars['csl_prop'])



### Scale Data

**Need to scale features for modeling purposes because of of the wide ranges of values and because PCA analysis requires scaling of features.**

from sklearn.preprocessing import StandardScaler
features= StandardScaler().fit_transform(dfClaims_quantVars)

### Principal Component Analysis

**Some of the variables are highly correlated to a great extent and there are still a great many features. Conduct PCA analysis for dimensionalty reduction and to address multicollinearity.**

#PCA
# https://www.datacamp.com/community/tutorials/principal-component-analysis-in-python
from sklearn.decomposition import PCA


pca=PCA(n_components=0.99, whiten=True)

features_pca = pca.fit_transform(features)

print("original number of features:", features.shape[1])

print("reduced number of features:", features_pca.shape[1])

# display the original feature names that compose the PCA components

model = PCA(n_components=16, whiten=True).fit(dfClaims_quantVars)
X_pc = model.transform(dfClaims_quantVars)

# number of components
n_pcs= model.components_.shape[0]

# get the index of the most important feature on EACH component
# LIST COMPREHENSION HERE
most_important = [np.abs(model.components_[i]).argmax() for i in range(n_pcs)]

initial_feature_names = ['months_as_customer', 'age', 'policy_deductable', 'policy_annual_premium', 'umbrella_limit', 'capital_gains', 'capital_loss', 'incident_hour_of_the_day', 'number_of_vehicles_involved', 'bodily_injuries', 'witnesses', 'total_claim_amount', 'injury_claim', 'property_claim', 'vehicle_claim', 'auto_year', 'csl_bodily', 'csl_prop']

# get the names
most_important_names = [initial_feature_names[most_important[i]] for i in range(n_pcs)]

# LIST COMPREHENSION HERE AGAIN
dic = {'PC{}'.format(i+1): most_important_names[i] for i in range(n_pcs)}

# build the dataframe
impPca = pd.DataFrame(sorted(dic.items()))


print(impPca)

# create new dataframe with PCA features
pcaDf = pd.DataFrame(data = features_pca, columns = ['PC1', 'PC2', 'PC3', 'PC4', 'PC5', 'PC6', 'PC7', 'PC8', 'PC9', 'PC10', 
                                                     'PC11', 'PC12', 'PC13', 'PC14', 'PC15', 'PC16'])


pcaDf.head()

pcaDf.shape

# add categorical variables back to df

pcaDf['policy_state'] = dfClaims['policy_state'] 
pcaDf['insured_sex'] = dfClaims['insured_sex']
pcaDf['insured_education_level'] = dfClaims['insured_education_level']
pcaDf['insured_occupation'] = dfClaims['insured_occupation']
pcaDf['insured_hobbies'] = dfClaims['insured_hobbies']
pcaDf['insured_relationship'] = dfClaims['insured_relationship']
pcaDf['incident_type'] = dfClaims['incident_type']
pcaDf['collision_type'] = dfClaims['collision_type'] 
pcaDf['incident_severity'] = dfClaims['incident_severity']
pcaDf['authorities_contacted'] = dfClaims['authorities_contacted']
pcaDf['incident_state'] = dfClaims['incident_state']
pcaDf['incident_city'] = dfClaims['incident_city']
pcaDf['property_damage'] = dfClaims['property_damage'] 
pcaDf['police_report_available'] = dfClaims['police_report_available']
pcaDf['auto_make'] = dfClaims['auto_make']
pcaDf['auto_model'] = dfClaims['auto_model']
pcaDf['fraud_reported'] = dfClaims['fraud_reported']

# check for missing values in PCA dataframe
for c in pcaDf.columns:
    miss = pcaDf[c].isnull().sum()
    if miss>0:
        print("{} has {} missing value(s)".format(c,miss))

pcaDf.shape

# the pcaDf now has a null row for some reason
is_NaN = pcaDf.isnull()
row_has_NaN = is_NaN.any(axis=1)
rows_with_NaN = pcaDf[row_has_NaN]

print(rows_with_NaN)

# remove the new null row
pcaDf =pcaDf.drop(pcaDf.index[290])

is_NaN = pcaDf.isnull()
row_has_NaN = is_NaN.any(axis=1)
rows_with_NaN = pcaDf[row_has_NaN]

print(rows_with_NaN)

# export the revised pca df to a csv file
pcaDf.to_csv('pcaClaimsLog.csv', index = False)

> Instead of running the above code multiple times, import the dataframes for subsequent analysis.

# import scaled, transformed and PCA df
pcaDF = pd.read_csv('pcaClaimsLog.csv')

# import cleaned df without re-running the cleaning code
df = pd.read_csv('dfClaims.csv')
