==> 3_Feature_Selection.ipynb <==
# Auto Insurance Claim Fraud Indicators and Classification

# Mary Donovan Martello

## The goal of this project was to identify significant features in fraudulent insurance claim transactions and to design predictive classification models to predict whether fraud was reported on the insurance claim transaction. This notebook includes feature selection.

# Part 3:  Feature Selection

# import data set and libraries for data preparation phase
import pandas as pd
import numpy as np
# importing regex module (search strings) RegEx can be used to check if a string contains the specified search pattern
import re


### The original dataset includes 1,000 prior claim transaction records. Each record has a mix of 38 quantitative and categorical data features about the claim filed, including information on the policy, insured, and automobile, aspects of the damage incident, and elements of the claim filed. The dataset also has a feature that indicates whether fraud was reported on each observation (i.e., either Y or N).

dfClaims = pd.read_csv('FradulentInsuranceClaims.csv')

dfClaims.head(2)

### The dataset in this notebook includes the original data that was cleaned in the 1_EDA_Prep notebook.

# import cleaned df without re-running the cleaning code
df = pd.read_csv('dfClaims.csv')

> # Feature Selection

> Prepare features to use in models.

# convert categorical data to numbers

#get the categorical data
cat_features = ['policy_state',
       'insured_sex', 'insured_education_level', 'insured_occupation',
       'insured_hobbies', 'insured_relationship', 'incident_type',
       'collision_type', 'incident_severity', 'authorities_contacted',
       'incident_state', 'incident_city', 'property_damage',
       'police_report_available', 'auto_make', 'auto_model']
df_cat = df[cat_features]

# One Hot Encoding 
dfDumm = pd.get_dummies(df_cat)

# combine the numerical features and the dummie features together
dfNum = df.drop(['policy_state', 'incident_location',
       'insured_sex', 'insured_education_level', 'insured_occupation',
       'insured_hobbies', 'insured_relationship', 'incident_type',
       'collision_type', 'incident_severity', 'authorities_contacted',
       'incident_state', 'incident_city', 'property_damage',
       'police_report_available', 'auto_make', 'auto_model', 'fraud_reported'], axis = 1)

Xdumm = pd.concat([dfNum, dfDumm], axis=1)

# create a whole target dataset that can be used for train and validation data splitting
y =  df['fraud_reported']
# for Random Forest when target is not label encoded
yRF = df['fraud_reported']

### Use CART classification and Random Forest Classifier to determine important features.

# https://machinelearningmastery.com/calculate-feature-importance-with-python/
# use CART classification feature importance
from sklearn.tree import DecisionTreeClassifier
from matplotlib import pyplot

# define the model
model = DecisionTreeClassifier()
# fit the model
model.fit(Xdumm, y)
# get importance
importance = model.feature_importances_
# summarize feature importance
for i,v in enumerate(importance):
	print('Feature: %0d, Score: %.5f' % (i,v))
# plot feature importance
plt.bar([x for x in range(len(importance))], importance)
plt.show()

# use random forest feature importance
# This approach can also be used with the bagging and extra trees algorithms.
# https://machinelearningmastery.com/calculate-feature-importance-with-python/
from sklearn.ensemble import RandomForestClassifier

# define the model
model = RandomForestClassifier()
# fit the model
model.fit(Xdumm, y)
# get importance
importance = model.feature_importances_
# summarize feature importance
for i,v in enumerate(importance):
	print('Feature: %0d, Score: %.5f' % (i,v))
# plot feature importance
plt.bar([x for x in range(len(importance))], importance)
plt.show()

#https://stackoverflow.com/questions/55466081/how-to-calculate-feature-importance-in-each-models-of-cross-validation-in-sklear

from sklearn.model_selection import cross_validate
from sklearn.ensemble import  RandomForestClassifier


clf=RandomForestClassifier(n_estimators =10, random_state = 42, class_weight="balanced")
output = cross_validate(clf, Xdumm, y, cv=2, scoring = 'accuracy', return_estimator =True)
for idx,estimator in enumerate(output['estimator']):
    print("Features sorted by their score for estimator {}:".format(idx))
    feature_importances = pd.DataFrame(estimator.feature_importances_,
                                       index = Xdumm.columns,
                                        columns=['importance']).sort_values('importance', ascending=False)
    print(feature_importances.head(20))
